+++ semcore/select/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangelog.com/en/1.0.0/).
 
-## [3.9.6] - 2023-06-22
+## [3.10.0] - 2023-06-23
+
+### Changed
+
+- Improved focus handling with `FilterTrigger`.
 
 ## [3.9.5] - 2023-06-16
 
@@ -11,8 +15,6 @@ CHANGELOG.md standards are inspired by [keepachangelog.com](https://keepachangel
 - Version patch update due to children dependencies update (`@semcore/dropdown-menu` [3.9.3 ~> 3.9.4]).
 
 ## [3.9.4] - 2023-06-15
-
-## [3.9.3] - 2023-06-13
 
 ### Fixed